refactor: bypass jsonrpc parameter facades internally
