+++ src/codex_a2a/jsonrpc/discovery_query.py
@@ -25,7 +25,7 @@ from codex_a2a.jsonrpc.errors import (
     upstream_http_error_response,
     upstream_unreachable_response,
 )
-from codex_a2a.jsonrpc.params import JsonRpcParamsValidationError
+from codex_a2a.jsonrpc.params_common import JsonRpcParamsValidationError
 
 if TYPE_CHECKING:
     from codex_a2a.jsonrpc.application import CodexSessionQueryJSONRPCApplication
