+++ src/codex_a2a/jsonrpc/review_control.py
@@ -9,8 +9,8 @@ from starlette.responses import Response
 
 from codex_a2a.jsonrpc.errors import invalid_params_response
 from codex_a2a.jsonrpc.owner_guard import validate_thread_owner
-from codex_a2a.jsonrpc.params import (
-    JsonRpcParamsValidationError,
+from codex_a2a.jsonrpc.params_common import JsonRpcParamsValidationError
+from codex_a2a.jsonrpc.review_control_params import (
     ReviewStartControlParams,
     ReviewWatchControlParams,
     parse_review_start_params,
