+++ src/codex_a2a/jsonrpc/session_control.py
@@ -22,18 +22,18 @@ from codex_a2a.jsonrpc.errors import (
     upstream_http_error_response,
     upstream_unreachable_response,
 )
-from codex_a2a.jsonrpc.params import (
+from codex_a2a.jsonrpc.params_common import JsonRpcParamsValidationError
+from codex_a2a.jsonrpc.payload_mapping import (
+    as_a2a_message,
+    message_to_item,
+)
+from codex_a2a.jsonrpc.session_control_params import (
     CommandControlParams,
-    JsonRpcParamsValidationError,
     PromptAsyncControlParams,
     ShellControlParams,
     parse_command_params,
     parse_prompt_async_params,
     parse_shell_params,
-)
-from codex_a2a.jsonrpc.payload_mapping import (
-    as_a2a_message,
-    message_to_item,
 )
 from codex_a2a.upstream.models import CodexRPCError, is_thread_not_found_error
