+++ src/codex_a2a/jsonrpc/thread_lifecycle_control.py
@@ -14,8 +14,8 @@ from codex_a2a.jsonrpc.errors import (
     upstream_unreachable_response,
 )
 from codex_a2a.jsonrpc.owner_guard import validate_thread_owner
-from codex_a2a.jsonrpc.params import (
-    JsonRpcParamsValidationError,
+from codex_a2a.jsonrpc.params_common import JsonRpcParamsValidationError
+from codex_a2a.jsonrpc.thread_lifecycle_params import (
     ThreadArchiveControlParams,
     ThreadForkControlParams,
     ThreadMetadataUpdateControlParams,
